# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Helm Glass application API
  version: 1.0.0
  description:
    Cookie-authenticated web cabinet REST API behind OAuth2 Proxy, with application
    authorization. MCP tools are served separately at /mcp and do not grant access to these WEB
    operations. GET /me establishes XSRF-TOKEN; web mutations send X-XSRF-TOKEN. Idempotency-Key is
    required only where declared; retain it when a mutation response is lost.
servers:
  - url: /api
security:
  - bearerAuth: []
  - webSession: []
paths:
...
  /me:
    get:
      operationId: me
      responses:
        "200":
          description: Current account; also establishes the XSRF-TOKEN cookie for web mutations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
    post:
      operationId: saveProfile
      description: Atomically updates the display name and optional portrait. Email remains read-only. Portraits are PNG/JPEG/WebP, at most 5 MiB and 16 megapixels; omitted avatar preserves the current photo.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [name, expectedVersion]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 300
                expectedVersion:
                  type: integer
                  format: int64
                  minimum: 0
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Saved profile or the receipt for an identical repeated operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "413":
          $ref: "#/components/responses/ApplicationError"
  /me/avatar:
    get:
      operationId: profileAvatar
      description: Original current portrait of the authenticated WEB account only; never a public file URL.
      parameters:
        - name: v
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Streamed original image with no-store and nosniff.
          content:
            image/png:
              schema: {type: string, format: binary}
            image/jpeg:
              schema: {type: string, format: binary}
            image/webp:
              schema: {type: string, format: binary}
        "404":
          $ref: "#/components/responses/ApplicationError"
  /tasks:
    get:
      operationId: tasks
      parameters:
        - $ref: "#/components/parameters/TaskIdFilter"
        - $ref: "#/components/parameters/Suggestions"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Site"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Direction"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Page<Task>; counters are returned by /tasks/summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskPage"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
    post:
      operationId: createTask
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Created draft or prepared task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
  /tasks/summary:
    get:
      operationId: taskSummary
      description: Uses all task list filters except status. Unknown is never represented as zero.
      responses:
        "200":
          description: "{active,succeeded,waitingForYou,total,sources:string[]}"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
      parameters:
        - $ref: "#/components/parameters/TaskIdFilter"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Site"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: task
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
    delete:
      operationId: deleteDraft
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: "Confirmed deletion: {deleted:true}; prepared tasks cannot be deleted individually."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deleted"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
  /tasks/{id}/commands:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      operationId: taskCommand
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCommand"
      responses:
        "200":
          description: Updated Task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
  /tasks/{id}/steps:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: taskSteps
      description: Durable business steps, newest declared first. Updates preserve identity and order. Page size is 10; beforeSequence fixes the viewed window, not step versions. Technical history is separate.
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Page"
        - name: beforeSequence
          in: query
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        "200":
          description: Registered business steps matching this query; total is not an estimated workload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStepPage"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
  /tasks/{id}/history:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: taskHistory
      description: Page size is always 10. Optional beforeSequence freezes the viewed history window.
      responses:
        "200":
          description: Page<{id,sequence,type,title,detail,createdAt}>
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Page"
        - name: type
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: OR over history types.
        - name: beforeSequence
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
  /tasks/{id}/result/rows:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: resultRows
      responses:
        "200":
          description: Page<{id,cells:object}>, server search/sort/page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultRowPage"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Direction"
        - name: sort
          in: query
          schema:
            type: string
          description: A published result column key; otherwise rows use their numeric id.
  /connections:
    get:
      operationId: connections
      responses:
        "200":
          description: Page<Connection>
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionPage"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Site"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Direction"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
    post:
      operationId: createConnection
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                site:
                  type: string
                  maxLength: 253
                startUrl:
                  type: string
                  maxLength: 4096
              description:
                A nonblank startUrl or site is required. With only site, the server uses
                https://{site}/; stored site is derived from the validated URL.
              anyOf:
                - required:
                    - startUrl
                - required:
                    - site
        required: true
      responses:
        "200":
          description: Connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
  /connections/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      operationId: connection
      responses:
        "200":
          description: Connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
    patch:
      operationId: renameConnection
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - expectedVersion
              properties:
                name:
                  type: string
                expectedVersion:
                  type: integer
        required: true
      responses:
        "200":
          description: Connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
    delete:
      operationId: deleteConnection
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: "Credentials deletion accepted: {deleted:true}; history and results remain."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deleted"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
...
  /browser-sessions/{id}/keep-open:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      operationId: keepBrowserOpen
      description: Extend the same live task browser by 15 minutes without resuming the task or changing control. A closing browser cannot be extended.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Browser deadline updated or protected while in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Browser"
        "403":
          $ref: "#/components/responses/ApplicationError"
        "404":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
...
  /notifications:
    get:
      operationId: notifications
      responses:
        "200":
          description: "{items:[{id,sequence,taskId,title,status,createdAt}],total,throughSequence}"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationInbox"
        "400":
          $ref: "#/components/responses/ApplicationError"
        "409":
          $ref: "#/components/responses/ApplicationError"
        "500":
          $ref: "#/components/responses/ApplicationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webSession:
      type: apiKey
      in: cookie
      name: _oauth2_proxy
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  schemas:
...
    Me:
      type: object
      required:
        - id
        - name
        - email
        - roles
        - status
        - version
        - avatarUrl
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
            enum:
              - USER
              - ADMIN
          maxItems: 2
        status:
          $ref: "#/components/schemas/AccountStatus"
        version:
          type: integer
          format: int64
          minimum: 0
        avatarUrl:
          type: [string, "null"]
          description: Authenticated relative URL of the current profile image, or null.
...
  responses:
    ApplicationError:
      description: Application error; see code and message. Validation failures may identify fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"